requirements.txt
cffi==1.16.0
dotenv-cli==3.3.0
cryptography==42.0.5
pycparser==2.22
PyJWT==2.8.0
